requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
colorama==0.4.4
python-telegram-bot==13.6
pytz==2021.1
six==1.16.0
tornado==6.1
